Fix type error in test data generation

+++ main.cc
@@ -16,7 +16,7 @@ int main(void) {
     float **X, *y;
 
     // Generate ground truth data
-    generate_test_data(X, y, [](int x) { 
+    generate_test_data(X, y, [](float x) { 
         return 2.5382 * cos(x)*x + x*x - 0.5; 
     });
